Converting to other mesh types

FerriteGridUtil.convert_to_makie_mesh — Method

convert_to_makie_mesh(grid; kwargs...)

Return a mesh that can be used for plotting with Makie.jl.

Keyword arguments:

displ: a Vector{Vec{dim, T}} containing the nodal displacements to plot a deformed grid. The default is undeformed.

cellset: to plot a subdomain of the grid, the default is the entire grid.

disconnectcells: to signify if the mesh will be used to plot a discontinous field. The default is false.

FerriteGridUtil.disconnect_field — Method

disconnect_field(field, grid; cellset)

Returns a vector of values that can be passed to a Makie.jl plotting function to plot a discontinous (across element boundaries) field.

Keyword arguments:

cellset: to plot a subdomain of the grid, the default is the entire grid.

mesh should be a mesh that was generated using the convert_to_makie_mesh(grid; disconnectcells=true) function. For constant values per cell field should be a Vector{Float64} with the length Ferrite.getncells(grid). Otherwise a Matrix can be passed with rows corresponding the the cellid::Int and the columns to the value of the nodes corresponding to that cell. Example matrix:

field = [
    0.44 0.29 # ... # data for nodes 1, 2 ... of element 1
    0.98 0.48 # ... # data for nodes 1, 2 ... of element 2
    0.32 0.55 # ... # data for nodes 1, 2 ... of element 3
]